스프링 시큐리티 적용/좋아요,댓글구현

+++ src/main/java/org/example/blog_project/post/PostService.java
@@ -4,16 +4,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.blog_project.member.Member;
 import org.example.blog_project.member.MemberRepository;
-import org.example.blog_project.post.dto.CreatePostResDto;
-import org.example.blog_project.post.dto.DetailPostDto;
-import org.example.blog_project.post.dto.PostForm;
-import org.example.blog_project.post.dto.PublishForm;
+import org.example.blog_project.post.dto.*;
 import org.example.blog_project.post_image.PostImageService;
 import org.example.blog_project.post_tag.PostTag;
 import org.example.blog_project.post_tag.PostTagRepository;
 import org.example.blog_project.post_tag.PostTagService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -27,10 +25,12 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Transactional(readOnly = true)
 public class PostService {
 
     @Value("${main.file.path}")
@@ -42,6 +42,7 @@ public class PostService {
     private final PostTagService postTagService;
     private final PostTagRepository postTagRepository;
 
+    @Transactional
     public CreatePostResDto createPost(Long memberId, PostForm postForm, List<MultipartFile> files){
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 유저"));
@@ -74,9 +75,13 @@ public class PostService {
                 .build();
     }
 
-    public String publishPost(PublishForm form, MultipartFile file) {
+    @Transactional
+    public String publishPost(PublishForm form, MultipartFile file,Long memberId) {
         Post post = postRepository.findById(form.getPostId())
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글"));
+        if(!memberId.equals(post.getMember().getMemberId())){
+            return "잘못된 사용자";
+        }
         post.setTemp(false);
 
         post.setIntroduce(form.getIntroduce());
@@ -102,10 +107,14 @@ public class PostService {
         return post.getPostUrl();
     }
 
+    @Transactional
+    public String updatePost(PostForm form,Long postId,List<MultipartFile> files, Long memberId){
 
-    public String updatePost(PostForm form,Long postId){
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글"));
+        if(!memberId.equals(post.getMember().getMemberId())){
+            return "잘못된 사용자";
+        }
 
         post.updateTitle(form.getTitle());
         post.updateContent(form.getContent());
@@ -113,18 +122,27 @@ public class PostService {
 
         postTagRepository.deleteAllByPost(post);
         post.getPostTagList().clear(); //arrayList 비우기
-
         if(form.getTags()!=null){
             postTagService.createPostTag(post, form.getTags());
+        }
+        try {
+            // 이미지 처리: PostImageService를 사용하여 이미지 업데이트
+            postImageService.updatePostImages(post, files);
+        }catch (IOException e){
+            log.info(e.getMessage());
         }
 
         postRepository.save(post);
         return "수정 성공";
     }
 
-    public String deletePost(Long postId){
+    @Transactional
+    public String deletePost(Long postId,Long memberId){
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글"));
+        if(!memberId.equals(post.getMember().getMemberId())){
+            return "잘못된 사용자";
+        }
 
         postTagService.deletePostTag(post);
         postRepository.delete(post);
@@ -132,6 +150,7 @@ public class PostService {
         return "삭제 성공";
     }
 
+    @Transactional
     public boolean deleteFile(Long postId, String fileName) {
         if (fileName == null || fileName.isEmpty()) {
             return false;
@@ -180,7 +199,10 @@ public class PostService {
     }
 
     public DetailPostDto getDetailPost(String loginId, String decodedTitle){
-        Post post = postRepository.findByPostUrl(decodedTitle)
+
+        String url = "/@" + loginId + "/" + decodedTitle;
+
+        Post post = postRepository.findByPostUrl(url)
                 .orElseThrow(() -> new RuntimeException("존재하지않는 게시글"));
         Optional<Post> previousPost = getPreviousPost(post.getMember().getMemberId(), post.getPostId());
         Optional<Post> nextPost = getNextPost(post.getMember().getMemberId(), post.getPostId());
@@ -192,8 +214,8 @@ public class PostService {
                 .series(post.getSeries())
                 .content(post.getContent())
                 .profileImageUrl(post.getMainImageUrl())
-                .prePostId(previousPost.map(Post::getPostId).orElse(null))
-                .nextPostId(nextPost.map(Post::getPostId).orElse(null))
+                .prePostUrl(previousPost.map(Post::getPostUrl).orElse(null))
+                .nextPostUrl(nextPost.map(Post::getPostUrl).orElse(null))
                 .build();
     }
 
@@ -206,5 +228,20 @@ public class PostService {
     public Optional<Post> getNextPost(Long memberId, Long postId) {
         List<Post> nextPosts = postRepository.findNextPost(memberId, postId);
         return nextPosts.isEmpty() ? Optional.empty() : Optional.of(nextPosts.get(0));
+    }
+
+    //TODO 모든 게시글 조회
+    public List<Post> getAllPosts(){
+        List<Post> allPosts = postRepository.findAll();
+        return null;
+    }
+    public List<TempPostDto> getAllTempPosts(){
+        List<Post> allByIsTemp = postRepository.findAllByIsTemp(true);
+        return allByIsTemp.stream().map(post -> new TempPostDto(
+                post.getTitle(),
+                post.getIntroduce(),
+                post.getCreatedAt(),
+                post.getMember().getName()
+        )).collect(Collectors.toList());
     }
 }
